backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
sqlalchemy
psycopg2-binary
python-dotenv
pyautogen
gunicorn
